Creating new group on frontend side
Adding the ability to create new group
